pyproject.toml
[tool.poetry]
name = "blaise-point-in-time-restore"
version = "0.1.0"
description = ""
authors = ["Jake West <jacob.west@ons.gov.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
pymysql = "^1.1.0"
cloud-sql-python-connector = "^1.4.2"
functions-framework = "^3.4.0"
requests = "^2.31.0"
google-auth = "^2.23.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
